Assignments for 11/1/2022
I messed around with some of the lines of code in section 5-1 to understand what exactly was happening. Once I understood what was happening, I switched the ellipses that were being drawn to stars and then used the given formula to draw 12 stars in a perfect circle. Then I colored the canvas blue to make it look like the EU flag.

In section5-2 I changed the star to a polygon with 9 points and sides, however I made it so that the user can manipulate the number of points / sides freely by making it an argument.

In section 5-q1, I didn't know what to do because there were no examples shows to me... Anyways, I made the rectangle red and then I added more arguments to the function balloon(). I added the color of the rectangle or balloon, the desired X coordinate, and the desired Y coordinate.